feat: checkbox toggle function
