+++ src/components/modals/playlist-modal/playlist-modal.jsx
@@ -1,61 +1,17 @@
 import styles from "./playlist-modal.module.css";
 import { useServices } from "../../../context/servicesContext/servicesContext";
-import {
-  addNewVideoToPlaylistOfUserService,
-  deleteVideoFromPlaylistOfUserService
-} from "../../../services";
-import { useData } from "../../../context/dataContext/dataContext";
-import { useToast } from "../../../custom-hooks/useToast";
-import { useAuth } from "../../../context/authContext/authenticationContext";
 
 const PlaylistModal = ({
   selectedVideo,
   showPlaylistModal,
   closePlaylistModal
 }) => {
-  const { state, dispatch } = useServices();
-  const { authToken } = useAuth();
-  const { videos } = useData();
-  const { showToast } = useToast();
+  const { state, addOrRemoveVideoFromPlaylist, videoExistsInThatPlaylist} = useServices();
 
   if (!showPlaylistModal) {
     return null;
   }
 
-  const videoExistsInThatPlaylist = (state.playlists.filter((item) => item._id === selectedVideo.id)) === undefined ? false : true;
-  const addOrRemoveVideoFromPlaylist = async ({ _id }) => {
-    // const videoExistsInThatPlaylist = (state.playlists.find((item) => item._id === _id)) === undefined ? false : true;
-    const video = videos.find((item) => item._id === selectedVideo.id);
-
-    try {
-      const response = videoExistsInThatPlaylist
-        ? ( videoExistsInThatPlaylist = false, await deleteVideoFromPlaylistOfUserService(
-            authToken,
-            state.playlists._id,
-            selectedVideo.id
-          ))
-        : ( videoExistsInThatPlaylist = true, await addNewVideoToPlaylistOfUserService(authToken, _id, video));
-      let singlePlaylist = state.playlists.map((playlist) => {
-        if (playlist._id === response.data.playlist._id) {
-          return response.data.playlist;
-        }
-        return playlist;
-      });
-      showToast("Video added to the playlist", "success");
-      dispatch({
-        type: "MANAGE_PLAYLIST",
-        payload: singlePlaylist
-      });
-    } catch (error) {
-      console.error(error);
-      showToast(
-        videoExistsInThatPlaylist
-          ? "Unable to delete video from playlist"
-          : "Unable to add video to playlist",
-        "error"
-      );
-    }
-  };
   return (
     <div className={styles.modal_wrapper}>
       <div className={styles.modal}>
@@ -70,6 +26,8 @@ const PlaylistModal = ({
         <div className={styles.modal_contents}>
           <ul className={styles.modal_content_list}>
             {state.playlists.map(({ title, _id }) => {
+              const currentPlaylist = state.playlists.find((item) => item._id === _id);
+              const videoExistsInThatPlaylist = currentPlaylist.videos.some((item) => item.id === selectedVideo.id);
               return (
                 <li className={styles.unordered_list} key={_id}>
                   <input
@@ -77,7 +35,7 @@ const PlaylistModal = ({
                     id={_id}
                     name={title}
                     checked={videoExistsInThatPlaylist}
-                    onChange={() => addOrRemoveVideoFromPlaylist({ _id })}
+                    onChange={() => addOrRemoveVideoFromPlaylist({ _id , selectedVideo })}
                   />
                   <label> {title} </label>
                 </li>
