+++ src/pages/MyPlaylistPage/MyPlaylistPage.jsx
@@ -21,8 +21,8 @@ const MyPlaylistPage = () => {
         {currentPlaylist.videos.map((element) => {
           return (
             <VideoCard
-              key={element._id}
-              id={element._id}
+              key={element.id}
+              id={element.id}
               title={element.title}
             />
           );
